requirements.txt
asgiref==3.2.4
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.8
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
python-decouple==3.3
pytz==2019.3
requests==2.23.0
responses==0.10.12
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
gunicorn==19.8.1
psycopg2==2.7.4
